Implement FRP-based room behavior system.
Also added a simple chat room behavior for testing purposes.

At the moment, a lot of this is just a huge mess. It will need some
clean up. I'm also considering making the FRP stuff a layer on top of
the normal room code, rather than having it integrated together like it
is now.