Use locks to ensure thread safety in rooms
Rooms are now locked during join and part operations to ensure thread
safety. This is probably the most elegant solution to the problem of
running IO actions on rooms concurrently, as only one thread should ever
be doing something like that at any given time.

No global mutexes were used in the making of this program.