Implement Reversi as a test case for the API.
This will certainly be split off into a separate library later on, but
I've added it for now as a test case to make sure the API can handle a
reasonable use case.

I've also added a "Players" module which provides a mechanism for
clients to join and play games.