Initial work for v0.4
Removed the use of ChanUUID for just UUIDs as Unique and UUID are significantly
different enough to not need the previous newtypes that used to be used in the
library.

Turned the old type synonyms into newtypes to better help with errors.

The project structure has been changed to be per library rather than classifying
each library arbitrarily.

Some nomenclature has been changed to hopefully reduce anologies that are used
when describing how to use the library. Modules are now services and requesters
are now clients. Perhaps this'll have to change again when writing the
documentation, but that depends on how fuzzy it is to describe the processes.